requirements.txt
beautifulsoup4>=4.11.0
selenium>=4.10.0
openai>=1.0.0
requests>=2.31.0
pandas>=2.0.0
undetected-chromedriver>=3.5.0
python-dotenv>=1.0.0
pyyaml>=6.0
pytesseract>=0.3.10
pdf2image>=1.16.0
streamlit>=1.28.0
plotly>=5.17.0

# API e Automação (para integração n8n)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

